Added glossary
A handy guide for someone who is considering installing this extension as it will let the person know that it does more than just replace the word "Millennial". It can be jarring when reading an article and the "time of shedding and cold rocks" or the "great ape-snake war"  is referenced and you have no idea why.